add user deletion from admin page, add user page make bet button

+++ src/main/java/lv/javaguru/java2/servlet/mvc/controllers/AdminPageController.java
@@ -34,7 +34,7 @@ public class AdminPageController implements MVCController {
     @Override
     public MVCModel processPost(HttpServletRequest req) {
         String deletedUserId = req.getParameter("deletedUserId");
-        if (deletedUserId != null) {
+        if ((deletedUserId != null)&&(!deletedUserId.equals(""))) {
             Long id = Long.parseLong(deletedUserId);
             adminService.deleteUserById(id);
         }
